Add saving location to file; Small fixes with map; Set up scale on the graphs
When we turn on tracking it will be automatically saved to file (one on
internal storage for app and one on external storage to have access).
Also set up to start and stop service when we click on switch.
And set up scale to be 1 min (60s) on graph (on X).